+++ pkg/analysis_server/lib/src/services/correction/dart/import_add_show.dart
@@ -30,7 +30,7 @@ class ImportAddShow extends CorrectionProducer {
       return;
     }
     // prepare whole import namespace
-    var importElement = importDirective.element2;
+    var importElement = importDirective.element;
     if (importElement == null) {
       return;
     }