+++ pkg/analyzer/lib/src/summary2/metadata_resolver.dart
@@ -101,7 +101,7 @@ class MetadataResolver extends ThrowingAstVisitor<void> {
     // We might have already accessed metadata flags, e.g. `hasDeprecated`,
     // before we finished metadata resolution, during `PrefixScope` building.
     // So, these flags are not accurate anymore, and we need to reset them.
-    node.element2!.resetMetadataFlags();
+    node.element!.resetMetadataFlags();
   }
 
   @override