+++ pkg/analyzer/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 5.2.0-dev
 * Deprecated `Element.enclosingElement3`, use `enclosingElement` instead.
+* Deprecated `Directive.element2`, use `element` instead.
 
 ## 5.1.0
 * Deprecated `AstNode.name2`, use `name` instead.